+++ src/commands/mod.rs
@@ -1,5 +1,6 @@
 pub mod index;
 pub mod info;
+pub mod init;
 pub mod install;
 pub mod list;
 pub mod outdated;